fix(core): align keycloak provisioning in one command
