+++ service/cmd/provisionKeycloak.go
@@ -2,10 +2,15 @@ package cmd
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"log/slog"
+	"os"
 
 	"github.com/opentdf/platform/lib/fixtures"
-	"github.com/opentdf/platform/service/internal/config"
 	"github.com/spf13/cobra"
+	"gopkg.in/yaml.v2"
 )
 
 const (
@@ -13,59 +18,118 @@ const (
 	provKcUsernameFlag = "username"
 	provKcPasswordFlag = "password"
 	provKcRealmFlag    = "realm"
+	provKcFilenameFlag = "file"
 	provKcInsecure     = "insecure"
 )
 
 var (
-	provisionKeycloakCmd = &cobra.Command{
-		Use:   "keycloak",
-		Short: "Run local provision of keycloak data",
-		Long: `
- ** Local Development and Testing Only **
- This command will create the following Keyclaok resource:
- - Realm
- - Roles
- - Client
- - Users
-
- This command is intended for local development and testing purposes only.
- `,
-		RunE: func(cmd *cobra.Command, _ []string) error {
-			kcEndpoint, _ := cmd.Flags().GetString(provKcEndpointFlag)
-			realmName, _ := cmd.Flags().GetString(provKcRealmFlag)
-			kcUsername, _ := cmd.Flags().GetString(provKcUsernameFlag)
-			kcPassword, _ := cmd.Flags().GetString(provKcPasswordFlag)
-			configFile, _ := cmd.Flags().GetString(configFileFlag)
-			configKey, _ := cmd.Flags().GetString(configKeyFlag)
-			insecure, _ := cmd.Flags().GetBool(provKcInsecure)
-
-			config, err := config.LoadConfig(configKey, configFile)
-			if err != nil {
-				return err
-			}
-
-			kcConnectParams := fixtures.KeycloakConnectParams{
-				BasePath:         kcEndpoint,
-				Username:         kcUsername,
-				Password:         kcPassword,
-				Realm:            realmName,
-				Audience:         config.Server.Auth.Audience,
-				AllowInsecureTLS: insecure,
-			}
-
-			return fixtures.SetupKeycloak(context.Background(), kcConnectParams)
-		},
-	}
+	provKeycloakFilename = "./service/cmd/keycloak_data.yaml"
+	keycloakData         fixtures.KeycloakData
 )
+
+var provisionKeycloakCmd = &cobra.Command{
+	Use:   "keycloak",
+	Short: "Run local provision of keycloak data",
+	Long: `
+	** Local Development and Testing Only **
+	This command will create the following Keyclaok resource:
+	- Realm
+	- Roles
+	- Client
+	- Users
+
+	This command is intended for local development and testing purposes only.
+	`,
+	RunE: func(cmd *cobra.Command, _ []string) error {
+		kcEndpoint, _ := cmd.Flags().GetString(provKcEndpointFlag)
+		kcUsername, _ := cmd.Flags().GetString(provKcUsernameFlag)
+		kcPassword, _ := cmd.Flags().GetString(provKcPasswordFlag)
+		keycloakFilename, _ := cmd.Flags().GetString(provKcFilenameFlag)
+
+		LoadKeycloakData(keycloakFilename)
+		ctx := context.Background()
+
+		kcParams := fixtures.KeycloakConnectParams{
+			BasePath:         kcEndpoint,
+			Username:         kcUsername,
+			Password:         kcPassword,
+			Realm:            "",
+			AllowInsecureTLS: true,
+		}
+
+		err := fixtures.SetupCustomKeycloak(ctx, kcParams, keycloakData)
+		if err != nil {
+			return err
+		}
+
+		return nil
+	},
+}
+
+var provisionKeycloakFromConfigCmd = &cobra.Command{
+	Use: "keycloak-from-config",
+	RunE: func(cmd *cobra.Command, args []string) error {
+		fmt.Println("Command keycloak-from-config has been deprecated. Please use command 'keycloak' instead.")
+		return nil
+	},
+}
+
+func convert(i interface{}) interface{} {
+	switch x := i.(type) {
+	case map[interface{}]interface{}:
+		m2 := map[string]interface{}{}
+		for k, v := range x {
+			m2[k.(string)] = convert(v) //nolint:forcetypeassert // allow type assert
+		}
+		return m2
+	case []interface{}:
+		for i, v := range x {
+			x[i] = convert(v)
+		}
+	}
+	return i
+}
+
+func LoadKeycloakData(file string) {
+	fmt.Println("file", file)
+	yamlData := make(map[interface{}]interface{})
+
+	f, err := os.Open(file)
+	if err != nil {
+		panic(fmt.Errorf("error when opening YAML file: %s", err.Error()))
+	}
+
+	fileData, err := io.ReadAll(f)
+	if err != nil {
+		panic(fmt.Errorf("error reading YAML file: %s", err.Error()))
+	}
+
+	err = yaml.Unmarshal(fileData, &yamlData)
+	if err != nil {
+		panic(fmt.Errorf("error unmarshaling yaml file %s", err.Error()))
+	}
+
+	cleanedYaml := convert(yamlData)
+
+	kcData, err := json.Marshal(cleanedYaml)
+	if err != nil {
+		panic(fmt.Errorf("error converting yaml to json: %s", err.Error()))
+	}
+
+	if err := json.Unmarshal(kcData, &keycloakData); err != nil {
+		slog.Error("could not unmarshal json into data object", slog.String("error", err.Error()))
+		panic(err)
+	}
+}
 
 func init() {
 	provisionKeycloakCmd.Flags().StringP(provKcEndpointFlag, "e", "http://localhost:8888/auth", "Keycloak endpoint")
 	provisionKeycloakCmd.Flags().StringP(provKcUsernameFlag, "u", "admin", "Keycloak username")
 	provisionKeycloakCmd.Flags().StringP(provKcPasswordFlag, "p", "changeme", "Keycloak password")
-	provisionKeycloakCmd.Flags().StringP(provKcRealmFlag, "r", "opentdf", "OpenTDF Keycloak Realm name")
-	provisionKeycloakCmd.Flags().BoolP(provKcInsecure, "", false, "Ignore tls verification when connecting to keycloak. --insecure to disable.")
+	provisionKeycloakCmd.Flags().StringP(provKcFilenameFlag, "f", provKeycloakFilename, "Keycloak config file")
 
 	provisionCmd.AddCommand(provisionKeycloakCmd)
 
-	rootCmd.AddCommand(provisionKeycloakCmd)
+	// Deprecated command
+	provisionCmd.AddCommand(provisionKeycloakFromConfigCmd)
 }
